[BREAKING CHANGE]: now LinkSampleUnifier returns WriterLoggingM monad.
this is because behavior of the unifier is complicated and often
counter-intuitive. We have to log it.

Although this is a breaking change, I don't think it's so serious. This
is because (1) the change in type is hidden by LinkSampleUnifier type
synonym and (2) the change in data (implementation) is usually hidden by
'unifyStd' function.