cli: Snapshot: BREAKING: now the parser returns a new opaque type.
This is because we want to support --duration option combined with
either --time-to or --time-from option. That kind of selective
combination of options is not allowed by optparse-applicative
alone. So, I changed the API design. Option Parser returns an
intermediate type that directly corresponds to CLI options, and then,
converts that type to Query. The conversion may fail for invalid
combination of options.